Fully remove the old resolve implementation
