+++ src/lib.rs
@@ -2,8 +2,7 @@ mod api;
 mod cache;
 mod cli;
 mod config;
-mod fetch;
-mod fetch2;
+mod engine;
 mod flock;
 mod git;
 mod model;
